requirements.txt
Flask
google-genai
moviepy
whisper
torch
coqui-tts
soundfile
noisereduce
librosa
numpy
Pillow
opencv-python
requests
gunicorn